requirements.txt

streamlit>=1.36.0
insightface>=0.7.3
onnxruntime>=1.18.0
opencv-python>=4.9.0.80
numpy>=1.26.4
Pillow>=10.3.0
scikit-learn>=1.4.2
tqdm>=4.66.4
geopy>=2.4.1
piexif>=1.1.3
